+++ src/main.js
@@ -3,6 +3,7 @@ import FilterView from './view/filter-view.js';
 import TripInfoView from './view/trip-info-view.js';
 import BoardPresenter from './presenter/board-presenter.js';
 import PointsModel from './model/points-model.js';
+import { generateFilter } from './mock/filter.js';
 
 const siteHeaderElement = document.querySelector('.trip-main');
 const siteFilterElement = document.querySelector('.trip-controls__filters');
@@ -15,7 +16,9 @@ const boardPresenter = new BoardPresenter(
     pointsModel
   });
 
+const filters = generateFilter(pointsModel.points);
+
 render(new TripInfoView(), siteHeaderElement, RenderPosition.AFTERBEGIN);
-render(new FilterView(), siteFilterElement);
+render(new FilterView({ filters }), siteFilterElement);
 
 boardPresenter.init();
